Improve project name detection
When called via docker, the effective app directory is not likely to be volume
mounted to a dir with the same name as the git repo. If the project lacks a
project.name file (JupiterOne internal convention), it will result in incorrect
project name at fact-discovery time.

Instead, try to pull the repo name from git origin URL config, as this is very likely to be set.