Pin to specific node version: 14.20.1
Certain versions of typescript, when used with the ttypescript wrapper, appear to be sensitive to node version changes.

The failing behavior is an error of the form:

```
$ yarn ttsc -p tsconfig-dist.json
$ /opt/jupiterone/node_modules/.bin/ttsc -p tsconfig-dist.json
/opt/jupiterone/node_modules/ttypescript/lib/loadTypescript.js:29
    var _e = ts.versionMajorMinor.split('.'), major = _e[0], minor = _e[1];
                                  ^

TypeError: Cannot read properties of undefined (reading 'split')
    at Object.loadTypeScript (/opt/jupiterone/node_modules/ttypescript/lib/loadTypescript.js:29:35)
```

This occurs with Typescript 4.3.1 and the latest point releases of node 12, 14, and 18. Node 14.20.1 works as expected. CI started failing when GitHub changed the point release of the `12.x` version spec under the hood. This change pins to a known-working version in GHA and docker.

Upgrading @jupiterone/typescript-tools was considered but will require changing basically all of the dev dependencies and it is unclear whether the `.d.ts` files provided by several of the prod dependencies (oclif, etc) are compatible with the newer `ttsc`.